feat: add `{ mode }` option, refactor pattern
