+++ packages/core/streams/anthropic-stream.ts
@@ -1,34 +1,13 @@
 import { AIStream } from './ai-stream'
-import { AIStreamCallbacks } from "./types"
-
-function parseAnthropicStream(): (data: string) => string | void {
-  let previous = ''
-
-  return data => {
-    const json = JSON.parse(data as string) as {
-      completion: string
-      stop: string | null
-      stop_reason: string | null
-      truncated: boolean
-      log_id: string
-      model: string
-      exception: string | null
-    }
-
-    // Anthropic's `completion` field is cumulative unlike OpenAI's
-    // deltas. In order to compute the delta, we must slice out the text
-    // we previously received.
-    const text = json.completion
-    const delta = text.slice(previous.length)
-    previous = text
-
-    return delta
-  }
-}
+import { TEXT_PARSERS } from './parsers'
+import { AIStreamCallbacks } from './types'
 
 export function AnthropicStream(
   res: Response,
-  cb?: AIStreamCallbacks
+  callbacks?: AIStreamCallbacks
 ): ReadableStream {
-  return AIStream(res, parseAnthropicStream(), cb)
+  return AIStream(res, {
+    parser: TEXT_PARSERS.anthropic(),
+    ...callbacks
+  })
 }
