+++ packages/core/streams/langchain-stream.ts
@@ -1,11 +1,11 @@
-import { createCallbacksTransformer } from './ai-stream'
-import { AIStreamCallbacks } from "./types"
+import { createCallbacksTransform } from './transforms'
+import { AIStreamCallbacks } from './types'
 
 export function LangChainStream(callbacks?: AIStreamCallbacks) {
   const stream = new TransformStream()
   const writer = stream.writable.getWriter()
   return {
-    stream: stream.readable.pipeThrough(createCallbacksTransformer(callbacks)),
+    stream: stream.readable.pipeThrough(createCallbacksTransform(callbacks)),
     handlers: {
       handleLLMNewToken: async (token: string) => {
         await writer.ready
