+++ packages/core/streams/types.ts
@@ -2,7 +2,7 @@
  * Whether the stream should parse the content from an API response or return
  * it unmodified.
  */
-export type AIStreamMode = "text" | "raw"
+export type AIStreamMode = 'text' | 'raw'
 
 /**
  * All of the configuration options for an AI stream.
@@ -18,13 +18,13 @@ export interface AIStreamCallbacks {
   onStart?: () => Promise<void>
   /**
    * Called when the stream is finished.
-   * 
+   *
    * @param completion The complete text response for this stream.
    */
   onCompletion?: (completion: string) => Promise<void>
   /**
    * Called when a token is received from the stream.
-   * 
+   *
    * @param token The token received.
    */
   onToken?: (token: string) => Promise<void>
